+++ src/filter.cpp
@@ -247,7 +247,8 @@ SEXP filter_not_grouped( DataFrame df, List args, const DataDots& dots){
 // [[Rcpp::export]]
 SEXP filter_impl( DataFrame df, List args, Environment env){
     check_valid_colnames(df) ;
-            
+    assert_all_white_list(df) ;
+    
     if( args.size() == 0 ) return df ;
     
     // special case